+++ src/TestableHttpClient/HttpRequestMessagesCheckExtensions.cs
@@ -10,7 +10,7 @@ public static class HttpRequestMessagesCheckExtensions
     /// <returns>The <seealso cref="IHttpRequestMessagesCheck"/> for further assertions.</returns>
     public static IHttpRequestMessagesCheck WithRequestUri(this IHttpRequestMessagesCheck check, string pattern) => WithRequestUri(check, pattern, null);
 
-    [Obsolete("Please use an overload without the 'ignoreCase', since ignoring casing is now controlled globally.")]
+    [Obsolete("Please use an overload without the 'ignoreCase', since ignoring casing is now controlled globally.", true)]
     public static IHttpRequestMessagesCheck WithRequestUri(this IHttpRequestMessagesCheck check, string pattern, bool ignoreCase) => WithRequestUri(check, pattern, ignoreCase, null);
 
     /// <summary>
@@ -22,7 +22,7 @@ public static class HttpRequestMessagesCheckExtensions
     /// <returns>The <seealso cref="IHttpRequestMessagesCheck"/> for further assertions.</returns>
     public static IHttpRequestMessagesCheck WithRequestUri(this IHttpRequestMessagesCheck check, string pattern, int expectedNumberOfRequests) => WithRequestUri(check, pattern, (int?)expectedNumberOfRequests);
 
-    [Obsolete("Please use an overload without the 'ignoreCase', since ignoring casing is now controlled globally.")]
+    [Obsolete("Please use an overload without the 'ignoreCase', since ignoring casing is now controlled globally.", true)]
     public static IHttpRequestMessagesCheck WithRequestUri(this IHttpRequestMessagesCheck check, string pattern, bool ignoreCase, int expectedNumberOfRequests) => WithRequestUri(check, pattern, ignoreCase, (int?)expectedNumberOfRequests);
 
     private static IHttpRequestMessagesCheck WithRequestUri(this IHttpRequestMessagesCheck check, string pattern, int? expectedNumberOfRequests)
@@ -48,36 +48,8 @@ public static class HttpRequestMessagesCheckExtensions
         return check.WithFilter(x => x.RequestUri is not null && uriPattern.Matches(x.RequestUri, check.Options.UriPatternMatchingOptions), expectedNumberOfRequests, condition);
     }
 
-    [Obsolete("Please use an overload without the 'ignoreCase', since ignoring casing is now controlled globally.")]
-    private static IHttpRequestMessagesCheck WithRequestUri(this IHttpRequestMessagesCheck check, string pattern, bool ignoreCase, int? expectedNumberOfRequests)
-    {
-        if (check == null)
-        {
-            throw new ArgumentNullException(nameof(check));
-        }
-
-        if (string.IsNullOrEmpty(pattern))
-        {
-            throw new ArgumentNullException(nameof(pattern));
-        }
-
-        var condition = string.Empty;
-        if (pattern != "*")
-        {
-            condition = $"uri pattern '{pattern}'";
-        }
-
-        UriPattern uriPattern = UriPatternParser.Parse(pattern);
-        var options = new UriPatternMatchingOptions
-        {
-            HostCaseInsensitive = ignoreCase,
-            PathCaseInsensitive = ignoreCase,
-            SchemeCaseInsensitive = ignoreCase,
-            QueryCaseInsensitive = ignoreCase
-        };
-
-        return check.WithFilter(x => x.RequestUri is not null && uriPattern.Matches(x.RequestUri, options), expectedNumberOfRequests, condition);
-    }
+    [Obsolete("Please use an overload without the 'ignoreCase', since ignoring casing is now controlled globally.", true)]
+    private static IHttpRequestMessagesCheck WithRequestUri(this IHttpRequestMessagesCheck check, string pattern, bool ignoreCase, int? expectedNumberOfRequests) => check;
 
     /// <summary>
     /// Asserts whether requests were made with a given querystring based on a pattern. For asserting the decoded version of the querystring is used.
@@ -85,7 +57,7 @@ public static class HttpRequestMessagesCheckExtensions
     /// <param name="check">The implementation that hold all the request messages.</param>
     /// <param name="pattern">The querystring pattern that is expected.</param>
     /// <returns>The <seealso cref="IHttpRequestMessagesCheck"/> for further assertions.</returns>
-    [Obsolete("Use WithRequestUri instead, since it now properly supports QueryStrings as well")]
+    [Obsolete("Use WithRequestUri instead, since it now properly supports QueryStrings as well", true)]
     public static IHttpRequestMessagesCheck WithQueryString(this IHttpRequestMessagesCheck check, string pattern) => WithQueryString(check, pattern, null);
 
     /// <summary>
@@ -95,31 +67,11 @@ public static class HttpRequestMessagesCheckExtensions
     /// <param name="pattern">The querystring pattern that is expected.</param>
     /// <param name="expectedNumberOfRequests">The expected number of requests.</param>
     /// <returns>The <seealso cref="IHttpRequestMessagesCheck"/> for further assertions.</returns>
-    [Obsolete("Use WithRequestUri instead, since it now properly supports QueryStrings as well")]
+    [Obsolete("Use WithRequestUri instead, since it now properly supports QueryStrings as well", true)]
     public static IHttpRequestMessagesCheck WithQueryString(this IHttpRequestMessagesCheck check, string pattern, int expectedNumberOfRequests) => WithQueryString(check, pattern, (int?)expectedNumberOfRequests);
 
-    [Obsolete("Use WithRequestUri instead, since it now properly supports QueryStrings as well")]
-    private static IHttpRequestMessagesCheck WithQueryString(this IHttpRequestMessagesCheck check, string pattern, int? expectedNumberOfRequests)
-    {
-        if (check == null)
-        {
-            throw new ArgumentNullException(nameof(check));
-        }
-
-        if (pattern == null)
-        {
-            throw new ArgumentNullException(nameof(pattern));
-        }
-
-        var condition = pattern switch
-        {
-            "" => "no querystring",
-            "*" => "any querystring",
-            _ => $"querystring pattern '{pattern}'"
-        };
-
-        return check.WithFilter(x => x.HasQueryString(pattern), expectedNumberOfRequests, condition);
-    }
+    [Obsolete("Use WithRequestUri instead, since it now properly supports QueryStrings as well", true)]
+    private static IHttpRequestMessagesCheck WithQueryString(this IHttpRequestMessagesCheck check, string pattern, int? expectedNumberOfRequests) => check;
 
     /// <summary>
     /// Asserts whether requests were made with a given HTTP Method.